backend/requirements.txt
matplotlib>=3.5.1
numpy>=1.22.2
pandas>=1.4.0
seaborn>=0.11.2
requests>=2.22.0
urllib3>=1.25.8

# web crawling
beautifulsoup4>=4.10.0

# django 
django>=4.0.3
django-allauth>=0.49.0
django-rest-auth>=0.9.5
djangorestframework>=3.13.1